Move isI18nEnabled and translate functions to a dedicated file
