+++ packages/core/src/internal/i18n-utils.ts
@@ -14,10 +14,22 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+import { GlobalConfig } from '../util/config';
+
 /**
- * @internal
+ * @private
  */
-export const doEval = (expression: string): any => {
-  // eslint-disable-next-line no-eval -- valid here as we want this function to be the only place in the codebase that uses eval
-  return eval(expression);
-};
+export function isI18nEnabled(): boolean {
+  return GlobalConfig.i18n.isEnabled();
+}
+
+/**
+ * @private
+ */
+export function translate(
+  key?: string | null,
+  params?: any[] | null,
+  defaultValue?: string
+): string | null {
+  return GlobalConfig.i18n.get(key, params, defaultValue);
+}
